chore: reorganisation of tests to enable as more parallel tests as it possible
