+++ src/test/java/net/serenitybdd/jbehave/WhenRunningWebJBehaveStoriesUsingGiven.java
@@ -1,0 +1,52 @@
+package net.serenitybdd.jbehave;
+
+import net.thucydides.core.model.TestOutcome;
+import net.thucydides.core.model.TestResult;
+import net.thucydides.core.model.TestStep;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.util.List;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+
+public class WhenRunningWebJBehaveStoriesUsingGiven extends AbstractJBehaveStory {
+
+    final static class AStorySample extends SerenityStories {
+        public AStorySample() {
+            super();
+        }
+
+        protected AStorySample(String storyName) {
+            findStoriesCalled(storyName);
+        }
+    }
+
+    @Before
+    public void reset_driver() {
+        environmentVariables.setProperty("webdriver.driver", "phantomjs");
+    }
+
+    @Test
+    public void browser_should_not_closed_between_given_stories_and_scenario_steps() throws Throwable {
+
+        // Given
+        SerenityStories story = newStory("aBehaviorWithGivenStories.story");
+
+        // When
+        run(story);
+
+        // Then
+        List<TestOutcome> outcomes = loadTestOutcomes();
+        assertThat(outcomes.get(0).getResult(), is(TestResult.SUCCESS));
+    }
+
+    @Test
+    public void two_scenarii_using_the_same_given_story_should_return_two_test_outcomes() throws Throwable {
+        SerenityStories story = newStory("LookupADefinitionSuite.story");
+        run(story);
+        List<TestOutcome> outcomes = loadTestOutcomes();
+        assertThat(outcomes.size(), is(2));
+    }
+}
